Hostel Database Management System
It is a java-Mysql database connectivity. I have created this project in the netbeans. Firstly, create database in your mysql. In the "login.java", you have to enter id and password as "admin". Then you will go to Home frame. Then you can add a student, add new rooms, add student in room. All the student details can be stored in DB. We can add info of left students. We can also find which student is leaving in which room. We can also store per month fees of student that he has given or not.